code to download soilGrids at multiple areas
This code uses XML to access the WCS publication of SoilGrids250m. It asks for your variables, depths, and quantiles of interest. You then need to define the bounding boxes for the areas you want to download. SoilGrids250m uses Homolsine projection, which is less common. For ease, this code takes bounding box coordinates in lat/long and reprojects them to homolosine.